requirements.txt
Pillow==2.5.1
wsgiref==0.1.2
